DDOA_API/requirements.txt
Flask    ==    2.0.2
Flask-SQLAlchemy ==   2.5.1
passlib      ==      1.7.4
python-decouple  ==   3.6
python-dotenv    ==  0.19.2
WTForms         ==    3.0.1
Flask-Migrate     ==  3.1.0
psycopg2-binary  ==   2.9.3
gunicorn==20.0.4
Flask-Mail       ==   0.9.1
Flask-WTF         ==  1.0.0
passlib           == 1.7.4
python-dotenv     ==  0.19.2
python-slugify    ==  5.0.2
email-validator==1.1.3
Flask-JWT       ==    0.3.2

